pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "helloworld"
authors = [{name = "Guillem Borrell", email = "borrell.guillem@bcg.com"}]
dynamic = ["version", "description"]
dependencies = ["azure-ai-ml", "azure-identity"]

[project.scripts]
helloworld-job = "helloworld.run:test_job"